Update predict.R
Add necessary logic for being called by visreg. It's a little ugly because predict() is called by visregResid, and visregPred 2x -- once with, and once without SEs. predict() needs to return a list when the SEs are called, but just a numeric vector if visregPred is called without them. Seems to work now, with a single line change to visreg